sweet alert and toastify added

+++ src/pages/MyList/MyList.jsx
@@ -4,6 +4,7 @@ import { FaEdit } from "react-icons/fa";
 import { MdDelete } from "react-icons/md";
 import myListBG from "../../images/myListBG.svg";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 const MyList = () => {
     const { user } = useContext(AuthContext);
     const [currentData, setCurrentData] = useState([]);
@@ -17,16 +18,40 @@ const MyList = () => {
     }, [])
 
     const handleDelete = id => {
-        fetch(`http://localhost:5000/all-tourists-spot/${id}`, {
-            method: 'DELETE',
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    const remainingSpots = currentData.filter(checkData => id !== checkData._id);
-                    setCurrentData(remainingSpots);
-                }
-            })
+        Swal.fire({
+            title: "Are you sure to delete it ?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/all-tourists-spot/${id}`, {
+                    method: 'DELETE',
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            const remainingSpots = currentData.filter(checkData => id !== checkData._id);
+                            setCurrentData(remainingSpots);
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            });
+                        }
+                    })
+            }
+            else {
+                Swal.fire({
+                    title: "Not Deleted!",
+                    text: "Your file has been safe.",
+                    icon: "error"
+                });
+            }
+        });
     }
 
     return (
